+++ crates/core/src/properties/temperature.rs
@@ -4,19 +4,20 @@ use crate::constants::BOLTZMANN;
 use crate::properties::energy::KineticEnergy;
 use crate::properties::IntrinsicProperty;
 use crate::system::System;
+use crate::internal::Float;
 
 /// Instantaneous temperature of the system.
 #[derive(Clone, Copy, Debug, Serialize, Deserialize)]
 pub struct Temperature;
 
 impl IntrinsicProperty for Temperature {
-    type Res = f64;
+    type Res = Float;
 
     fn calculate_intrinsic(&self, system: &System) -> <Self as IntrinsicProperty>::Res {
         let kinetic = KineticEnergy.calculate_intrinsic(system);
         // NOTE: Calculating DOF this way is a potentially nasty bug if future
         // support is added for degrees of freedom beyond just 3D particles.
-        let dof = (system.size() * 3) as f64;
+        let dof = (system.size() * 3) as Float;
         2.0 * kinetic / (dof * BOLTZMANN)
     }
 }
