+++ crates/core/src/thermostats.rs
@@ -6,6 +6,7 @@ use serde::{Deserialize, Serialize};
 use crate::properties::temperature::Temperature;
 use crate::properties::IntrinsicProperty;
 use crate::system::System;
+use crate::internal::Float;
 
 /// An algorithm used to control simulation temperature.
 #[typetag::serde(tag = "type")]
@@ -28,8 +29,8 @@ impl Thermostat for NullThermostat {}
 /// Berendsen weak coupling thermostat.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Berendsen {
-    target: f64,
-    tau: f64,
+    target: Float,
+    tau: Float,
 }
 
 impl Berendsen {
@@ -39,7 +40,7 @@ impl Berendsen {
     ///
     /// * `target` - Target temperature (Kelvin)
     /// * `tau` - Timestep of the thermostat expressed as a multiple of the integrator's timestep
-    pub fn new(target: f64, tau: f64) -> Berendsen {
+    pub fn new(target: Float, tau: Float) -> Berendsen {
         Berendsen { target, tau }
     }
 }
@@ -48,26 +49,26 @@ impl Berendsen {
 impl Thermostat for Berendsen {
     fn post_integrate(&mut self, system: &mut System) {
         let temperature = Temperature.calculate_intrinsic(system);
-        let factor = f64::sqrt(1.0 + (self.target / temperature - 1.0) / self.tau);
+        let factor = Float::sqrt(1.0 + (self.target / temperature - 1.0) / self.tau);
 
         // !!! this block is more efficient without `par_iter`
         system.velocities = system
             .velocities
             .iter()
             .map(|&v| v * factor)
-            .collect::<Vec<Vector3<f64>>>();
+            .collect::<Vec<Vector3<Float>>>();
     }
 }
 
 /// Nose-Hoover style thermostat.
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct NoseHoover {
-    target: f64,
-    freq: f64,
-    timestep: f64,
-    psi: f64,
-    factor: f64,
-    temperature: f64,
+    target: Float,
+    freq: Float,
+    timestep: Float,
+    psi: Float,
+    factor: Float,
+    temperature: Float,
 }
 
 impl NoseHoover {
@@ -78,14 +79,14 @@ impl NoseHoover {
     /// * `target` - Target temperature (Kelvin)
     /// * `freq` - Damping frequency
     /// * `timestep` - Simulation timestep
-    pub fn new(target: f64, freq: f64, timestep: f64) -> NoseHoover {
+    pub fn new(target: Float, freq: Float, timestep: Float) -> NoseHoover {
         NoseHoover {
             target,
             freq,
             timestep,
-            psi: 0 as f64,
-            factor: 0 as f64,
-            temperature: 0 as f64,
+            psi: 0 as Float,
+            factor: 0 as Float,
+            temperature: 0 as Float,
         }
     }
 }
@@ -100,14 +101,14 @@ impl Thermostat for NoseHoover {
         let dt = self.timestep;
         let psidot = self.freq.powi(2) * ((self.temperature / self.target) - 1.0);
         self.psi += psidot * (dt / 2.0);
-        self.factor = f64::exp(-self.psi * (dt / 2.0));
+        self.factor = Float::exp(-self.psi * (dt / 2.0));
 
         // !!! this block is more efficient without `par_iter`
         system.velocities = system
             .velocities
             .iter()
             .map(|&v| v * self.factor)
-            .collect::<Vec<Vector3<f64>>>();
+            .collect::<Vec<Vector3<Float>>>();
     }
 
     fn post_integrate(&mut self, system: &mut System) {
